реализована перезагрузка страниц
